add new parameter only-filenames for java format
If enabled, no formatting is executed unless there are explicitly passed
filenames. This enables to make the formatting conditional in multi
language project. That is if java files are not touched, no attempt is
made - which make it possible to commit without installing java.

Upgraded black due to https://github.com/psf/black/issues/2964